info command & exec & server
