+++ packages/cli/src/commands/server/action.ts
@@ -2,11 +2,7 @@ import createLogger, { type Logger } from "@lage-run/logger";
 import type { ReporterInitOptions } from "../../types/ReporterInitOptions.js";
 import { initializeReporters } from "../initializeReporters.js";
 import { createLageService } from "./lageService.js";
-import type { Command } from "commander";
-import type { LageClient } from "@lage-run/rpc";
-import { filterArgsForTasks } from "../run/filterArgsForTasks.js";
-import { ConnectError, createClient, createServer } from "@lage-run/rpc";
-import { simulateFileAccess } from "./simulateFileAccess.js";
+import { createServer } from "@lage-run/rpc";
 
 interface WorkerOptions extends ReporterInitOptions {
   nodeArg?: string[];
@@ -16,56 +12,7 @@ interface WorkerOptions extends ReporterInitOptions {
   shutdown: boolean;
 }
 
-async function tryCreateClient(host: string, port: number) {
-  const client = createClient({
-    baseUrl: `http://${host}:${port}`,
-    httpVersion: "1.1",
-  });
-
-  try {
-    const success = await client.ping({});
-    if (success.pong) {
-      return client;
-    }
-  } catch (e) {
-    if (e instanceof ConnectError) {
-      return undefined;
-    }
-
-    throw e;
-  }
-
-  return undefined;
-}
-
-async function executeOnServer(args: string[], client: LageClient, logger: Logger) {
-  const task = args.length === 1 ? args[0] : args[1];
-  const packageName = args.length > 1 ? args[0] : undefined;
-
-  if (!task) {
-    throw new Error("No task provided");
-  }
-
-  const { taskArgs } = filterArgsForTasks(args ?? []);
-
-  const response = await client.runTarget({
-    packageName,
-    task,
-    taskArgs,
-  });
-
-  logger.info(`Task ${response.packageName} ${response.task} exited with code ${response.exitCode} `);
-
-  process.exitCode = response.exitCode;
-
-  if (response.exitCode === 0) {
-    await simulateFileAccess(logger, response.inputs, response.outputs);
-  }
-
-  logger.info("Task execution finished");
-}
-
-export async function serverAction(options: WorkerOptions, command: Command) {
+export async function serverAction(options: WorkerOptions) {
   const { port = 5332, host = "localhost", timeout = 1 } = options;
 
   const logger = createLogger();
@@ -73,42 +20,24 @@ export async function serverAction(options: WorkerOptions, command: Command) {
   options.reporter = options.reporter ?? "json";
   initializeReporters(logger, options);
 
-  const client = await tryCreateClient(host, port);
+  logger.info(`Starting server on http://${host}:${port}`);
 
-  if (client) {
-    logger.info(`Executing on server http://${host}:${port}`);
+  const abortController = new AbortController();
 
-    const args = command.args;
-    await executeOnServer(args, client, logger);
-  } else {
-    logger.info(`Starting server on http://${host}:${port}`);
+  const lageService = await createLageService({
+    cwd: process.cwd(),
+    serverControls: {
+      abortController,
+      countdownToShutdown: () => resetTimer(logger, timeout, abortController, server),
+      clearCountdown: clearTimer,
+    },
+    logger,
+    maxWorkers: options.concurrency,
+  });
+  const server = await createServer(lageService, abortController);
 
-    const abortController = new AbortController();
-
-    const lageService = await createLageService({
-      cwd: process.cwd(),
-      serverControls: {
-        abortController,
-        countdownToShutdown: () => resetTimer(logger, timeout, abortController, server),
-        clearCountdown: clearTimer,
-      },
-      logger,
-      maxWorkers: options.concurrency,
-    });
-    const server = await createServer(lageService, abortController);
-
-    await server.listen({ host, port });
-    logger.info(`Server listening on http://${host}:${port}, timeout in ${timeout} seconds`);
-
-    const client = await tryCreateClient(host, port);
-
-    if (!client) {
-      throw new Error("Server could not be reached");
-    }
-
-    const args = command.args;
-    await executeOnServer(args, client, logger);
-  }
+  await server.listen({ host, port });
+  logger.info(`Server listening on http://${host}:${port}, timeout in ${timeout} seconds`);
 }
 
 let timeoutHandle: NodeJS.Timeout | undefined;
